refactor: 점수 계산, 승자 구하는 책임 분리
