+++ src/main/java/domain/piece/Side.java
@@ -5,17 +5,6 @@ public enum Side {
     WHITE,
     NEUTRAL;
 
-    public static Side calculateWinner(double whiteScore, double blackScore) {
-        if (whiteScore == blackScore) {
-            return Side.NEUTRAL;
-        }
-        if (whiteScore > blackScore) {
-            return Side.WHITE;
-        }
-
-        return Side.BLACK;
-    }
-
     public Side nextSide() {
         if (this == WHITE) {
             return BLACK;
